Make diagnostics configuarable
Currently, ts-jest is checking the type errors for source codes and all
their dependencies if "enableTsDiagnostics" is set as true. This makes
it very slow and I don't think that's what we want since the src code
type errors will be detected during the build time.

What we need should be a way to report the type errors only inside the
test files, because usually test codes(describe, test) will be excluded
when build.

Make the "enableTsDiagnostics" configuarable with a regex, so that user
can make diagnostics only happens on test files or any other files they
want.