fix(compiler): make sure language service use updated compiler options
When a file initially exists in compiler file set when `TsCompiler` is created, later in `getCompiledOutput` the value of `compilerOptions` which makes `LanguageService` is not aware of because `_updateMemoryCache` checks and sees that the file is the same so `projectVersion` value is not increased which makes `LanguageService` uses the outdated `_compilerOptions`.

This commit adds a comparison via JSON stringify between `_compilerOptions` and `_initialCompilerOptions`. If there is a difference, increase `projectVersion` for `LanguageService`.