fix: introduce `transpilation` option to replace `isolatedModules` option
When `isolatedModules` option was introduced, the purpose was to indicate that it is a way to enable "transpilation mode", which is similar to other tools like `Babel`, `swc`, `esbuild` are doing.

However, this created a big misunderstanding and confusion between `isolatedModules` option in `tsconfig` vs `ts-jest` option. Therefore, `ts-jest` renamed `isolatedModules` option to `transpilation` to make things clearer